move fork_into_arbiter to lib tests
